requirements.txt
typer
requests
colorama
shellingham
python-dotenv